add functionality to delate tables with customImageDialog

+++ reservations/domain/src/main/java/com/example/domain/model/Table.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "tables")
 data class Table(
     val shape: String,
-    @PrimaryKey(autoGenerate = false) val id: Long,
+    @PrimaryKey(autoGenerate = false) val id: Int,
 )
